requirements-minimal.txt
flask>=2.3.0
flask-cors>=4.0.0
streamlit>=1.20.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.31.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
stripe>=6.0.0
colorama>=0.4.6

# Heavy ML packages - can be added later
# torch>=2.2.0
# torchvision>=0.17.0
# transformers>=4.30.0 